gateway/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-consul
requests